+++ src/style/index.ts
@@ -1,69 +1,71 @@
-// vue devtool
-export const VUE_DEEP_CYAN = "#41b883";
-export const VUE_BLUE_GRAY = "#35495e";
+export default {
+  // vue devtool
+  VUE_DEEP_CYAN: "#41b883",
+  VUE_BLUE_GRAY: "#35495e",
 
-// arco design
-export const ARCO_DEEP_RED = "#F53F3F";
-export const ARCO_LIGHT_RED = "#F76560";
-export const ARCO_DEEP_ORANGE = "#FF7D00";
-export const ARCO_LIGHT_ORANGE = "#F77234";
-export const ARCO_DEEP_YELLOW = "#F7BA1E";
-export const ARCO_LIGHT_YELLOW = "#FADC19";
-export const ARCO_DEEP_GREEN = "#00B42A";
-export const ARCO_LIGHT_GREEN = "#9FDB1D";
-export const ARCO_DEEP_CYAN = "#14C9C9";
-export const ARCO_LIGHT_CYAN = "#37D4CF";
-export const ARCO_DEEP_BLUE = "#165DFF";
-export const ARCO_LIGHT_BLUE = "#3491FA";
-export const ARCO_DEEP_PURPLE = "#722ED1";
-export const ARCO_LIGHT_PURPLE = "#8D4EDA";
+  // arco design
+  ARCO_DEEP_RED: "#F53F3F",
+  ARCO_LIGHT_RED: "#F76560",
+  ARCO_DEEP_ORANGE: "#FF7D00",
+  ARCO_LIGHT_ORANGE: "#F77234",
+  ARCO_DEEP_YELLOW: "#F7BA1E",
+  ARCO_LIGHT_YELLOW: "#FADC19",
+  ARCO_DEEP_GREEN: "#00B42A",
+  ARCO_LIGHT_GREEN: "#9FDB1D",
+  ARCO_DEEP_CYAN: "#14C9C9",
+  ARCO_LIGHT_CYAN: "#37D4CF",
+  ARCO_DEEP_BLUE: "#165DFF",
+  ARCO_LIGHT_BLUE: "#3491FA",
+  ARCO_DEEP_PURPLE: "#722ED1",
+  ARCO_LIGHT_PURPLE: "#8D4EDA",
 
-// material ui
-export const MATERIAL_BLUE = "#2196f3";
-export const MATERIAL_GREY = "#9e9e9e";
-export const MATERIAL_RED = "#f44336";
-export const MATERIAL_YELLOW = "#ffeb3b";
-export const MATERIAL_GREEN = "#4caf50";
-export const MATERIAL_ORANGE = "#ff9800";
-export const MATERIAL_PURPLE = "#9c27b0";
-export const MATERIAL_TEAL = "#009688";
-export const MATERIAL_PINK = "#e91e63";
-export const MATERIAL_BROWN = "#795548";
-export const MATERIAL_CYAN = "#00bcd4";
-export const MATERIAL_LIME = "#cddc39";
-export const MATERIAL_DEEP_ORANGE = "#ff5722";
-export const MATERIAL_LIGHT_BLUE = "#03a9f4";
-export const MATERIAL_AMBER = "#ffc107";
-export const MATERIAL_INDIGO = "#3f51b5";
-export const MATERIAL_LIGHT_GREEN = "#8bc34a";
-export const MATERIAL_DEEP_PURPLE = "#673ab7";
-export const MATERIAL_YELLOW_GREEN = "#9ccc65";
-export const MATERIAL_DEEP_CYAN = "#006064";
-export const MATERIAL_DEEP_GREEN = "#2e7d32";
-export const MATERIAL_LIGHT_YELLOW = "#fff59d";
-export const MATERIAL_LIGHT_RED = "#e57373";
-export const MATERIAL_LIGHT_GREEN_2 = "#81c784";
-export const MATERIAL_LIGHT_ORANGE = "#ffcc80";
-export const MATERIAL_LIGHT_PURPLE = "#ba68c8";
-export const MATERIAL_LIGHT_GREY = "#bdbdbd";
-export const MATERIAL_LIGHT_PINK = "#f48fb1";
-export const MATERIAL_LIGHT_TEAL = "#4db6ac";
-export const MATERIAL_DARK_BLUE = "#0d47a1";
-export const MATERIAL_DARK_GREEN = "#1b5e20";
-export const MATERIAL_DARK_PINK = "#880e4f";
-export const MATERIAL_DARK_CYAN = "#00838f";
-export const MATERIAL_DARK_YELLOW = "#f9a825";
-export const MATERIAL_DARK_RED = "#b71c1c";
-export const MATERIAL_DARK_ORANGE = "#e65100";
-export const MATERIAL_DARK_GREY = "#616161";
-export const MATERIAL_DARK_PURPLE = "#4a148c";
-export const MATERIAL_DARK_TEAL = "#004d40";
-export const MATERIAL_DARK_LIME = "#827717";
-export const MATERIAL_DARK_AMBER = "#ff6f00";
-export const MATERIAL_DARK_INDIGO = "#1a237e";
-export const MATERIAL_DARK_BROWN = "#3e2723";
-export const MATERIAL_DARK_YELLOW_GREEN = "#689f38";
-export const MATERIAL_DARK_LIGHT_BLUE = "#01579b";
-export const MATERIAL_DARK_LIGHT_GREEN = "#33691e";
-export const MATERIAL_DARK_LIGHT_PINK = "#ad1457";
-export const MATERIAL_DARK_LIGHT_PURPLE = "#6a1b9a";
+  // material ui
+  MATERIAL_BLUE: "#2196f3",
+  MATERIAL_GREY: "#9e9e9e",
+  MATERIAL_RED: "#f44336",
+  MATERIAL_YELLOW: "#ffeb3b",
+  MATERIAL_GREEN: "#4caf50",
+  MATERIAL_ORANGE: "#ff9800",
+  MATERIAL_PURPLE: "#9c27b0",
+  MATERIAL_TEAL: "#009688",
+  MATERIAL_PINK: "#e91e63",
+  MATERIAL_BROWN: "#795548",
+  MATERIAL_CYAN: "#00bcd4",
+  MATERIAL_LIME: "#cddc39",
+  MATERIAL_DEEP_ORANGE: "#ff5722",
+  MATERIAL_LIGHT_BLUE: "#03a9f4",
+  MATERIAL_AMBER: "#ffc107",
+  MATERIAL_INDIGO: "#3f51b5",
+  MATERIAL_LIGHT_GREEN: "#8bc34a",
+  MATERIAL_DEEP_PURPLE: "#673ab7",
+  MATERIAL_YELLOW_GREEN: "#9ccc65",
+  MATERIAL_DEEP_CYAN: "#006064",
+  MATERIAL_DEEP_GREEN: "#2e7d32",
+  MATERIAL_LIGHT_YELLOW: "#fff59d",
+  MATERIAL_LIGHT_RED: "#e57373",
+  MATERIAL_LIGHT_GREEN_2: "#81c784",
+  MATERIAL_LIGHT_ORANGE: "#ffcc80",
+  MATERIAL_LIGHT_PURPLE: "#ba68c8",
+  MATERIAL_LIGHT_GREY: "#bdbdbd",
+  MATERIAL_LIGHT_PINK: "#f48fb1",
+  MATERIAL_LIGHT_TEAL: "#4db6ac",
+  MATERIAL_DARK_BLUE: "#0d47a1",
+  MATERIAL_DARK_GREEN: "#1b5e20",
+  MATERIAL_DARK_PINK: "#880e4f",
+  MATERIAL_DARK_CYAN: "#00838f",
+  MATERIAL_DARK_YELLOW: "#f9a825",
+  MATERIAL_DARK_RED: "#b71c1c",
+  MATERIAL_DARK_ORANGE: "#e65100",
+  MATERIAL_DARK_GREY: "#616161",
+  MATERIAL_DARK_PURPLE: "#4a148c",
+  MATERIAL_DARK_TEAL: "#004d40",
+  MATERIAL_DARK_LIME: "#827717",
+  MATERIAL_DARK_AMBER: "#ff6f00",
+  MATERIAL_DARK_INDIGO: "#1a237e",
+  MATERIAL_DARK_BROWN: "#3e2723",
+  MATERIAL_DARK_YELLOW_GREEN: "#689f38",
+  MATERIAL_DARK_LIGHT_BLUE: "#01579b",
+  MATERIAL_DARK_LIGHT_GREEN: "#33691e",
+  MATERIAL_DARK_LIGHT_PINK: "#ad1457",
+  MATERIAL_DARK_LIGHT_PURPLE: "#6a1b9a",
+};