+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+## 0.0.9 (2023-06-30)
+
+
+### Bug Fixes
+
+* 修复 styleValue 和 emoji 缺少类型提示的问题 ([9a2b613](https://github.com/Redstone-1/console-log-button/commit/9a2b6134d3bcc2250e09bd9fb2cab5eb462cfa76))
+* 修改 buttonLogUtils 导入导出, 完善其 ts 类型提示 ([4045cfb](https://github.com/Redstone-1/console-log-button/commit/4045cfb83dde61386f761b20830883132e9f97e1))
+
+
+### Features
+
+* 1. 增加代码压缩, 2. 增加 .d.ts 声明文件, 3. package.json 增加部分信息 ([11e71f8](https://github.com/Redstone-1/console-log-button/commit/11e71f8db28346f4b84d42bb3002b2717a45ea84))
+* 拆分 baseButtonLog 及 buttonLogUtils, 拆分 ts 类型文件, 新增 getDoubleButtonConfigs & getMaterialConfigs 方法, 简化代码 ([0879d65](https://github.com/Redstone-1/console-log-button/commit/0879d65958f676dae05799973f7317bd174c56c2))
+* 完成基本功能和项目基础 ([0feb411](https://github.com/Redstone-1/console-log-button/commit/0feb411d2f8e7495382c4a7f9a568de4970dc3c3))
+* 新增 getMaterialGhostConfigs 及 arco design 样式方法 ([5315fe2](https://github.com/Redstone-1/console-log-button/commit/5315fe27f7d2aae48acfb7287433ec13351e08d7))
+
+
+
 ## 0.0.8 (2023-06-29)